--- apps/landing-page/index.html
@@ -1,14 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-	<head>
-		<meta charset="UTF-8" />
-		<link rel="icon" type="image/svg+xml" href="/src/favicon.svg" />
-		<meta name="viewport" content="width=device-width, initial-scale=1.0" />
-		<title>Embrace Community</title>
-	</head>
-	<body class="font-inter antialiased bg-gradient-background bg-fixed text-gray-800 tracking-tight">
-		<noscript>You need to enable JavaScript to run this app.</noscript>
-		<div id="root"></div>
-		<script type="module" src="/src/main.jsx"></script>
-	</body>
-</html>
