--- apps/landing-page/package.json
@@ -1,24 +1,0 @@
-{
-	"name": "landing-page",
-	"version": "0.1.0",
-	"scripts": {
-		"dev": "vite",
-		"build": "vite build",
-		"preview": "vite preview"
-	},
-	"dependencies": {
-		"@tailwindcss/forms": "^0.5.3",
-		"aos": "^3.0.0-beta.6",
-		"react": "^18.2.0",
-		"react-dom": "^18.2.0",
-		"react-router-dom": "^6.8.1",
-		"react-transition-group": "^4.4.5"
-	},
-	"devDependencies": {
-		"@vitejs/plugin-react": "^3.1.0",
-		"autoprefixer": "^10.4.13",
-		"postcss": "^8.4.21",
-		"tailwindcss": "^3.2.6",
-		"vite": "^4.1.1"
-	}
-}
