--- apps/landing-page/src/css/tailwind.config.js
@@ -1,50 +1,0 @@
-module.exports = {
-	content: ["./index.html", "./src/**/*.{js,jsx,ts,tsx}"],
-	theme: {
-		extend: {
-			backgroundImage: {
-				"gradient-background": "url('/imgs/gradient-background.png')"
-			},
-			colors: {},
-			spacing: {
-				"9/16": "56.25%",
-				"3/4": "75%",
-				"1/1": "100%"
-			},
-			fontFamily: {
-				inter: ["Inter", "sans-serif"],
-				"architects-daughter": ['"Architects Daughter"', "sans-serif"]
-			},
-			fontSize: {
-				xs: "0.75rem",
-				sm: "0.875rem",
-				base: "1rem",
-				lg: "1.125rem",
-				xl: "1.25rem",
-				"2xl": "1.5rem",
-				"3xl": "2rem",
-				"4xl": "2.5rem",
-				"5xl": "3.25rem",
-				"6xl": "4rem"
-			},
-			inset: {
-				full: "100%"
-			},
-			letterSpacing: {
-				tighter: "-0.02em",
-				tight: "-0.01em",
-				normal: "0",
-				wide: "0.01em",
-				wider: "0.02em",
-				widest: "0.4em"
-			},
-			minWidth: {
-				10: "2.5rem"
-			},
-			scale: {
-				98: ".98"
-			}
-		}
-	},
-	plugins: [require("@tailwindcss/forms")]
-};
