--- apps/landing-page/vite.config.js
@@ -1,29 +1,0 @@
-import { defineConfig } from 'vite'
-import postcss from './postcss.config.js'
-import react from '@vitejs/plugin-react'
-
-// https://vitejs.dev/config/
-export default defineConfig({
-  define: {
-    'process.env': process.env
-  },
-  css: {
-    postcss,
-  },
-  plugins: [react()],
-  resolve: {
-    alias: [
-      {
-        find: /^~.+/,
-        replacement: (val) => {
-          return val.replace(/^~/, "");
-        },
-      },
-    ],
-  },
-  build: {
-    commonjsOptions: {
-      transformMixedEsModules: true,
-    }
-  } 
-})
