+++ apps/landing/.prettierrc.json
@@ -1,0 +1,4 @@
+{
+	"semi": true,
+	"trailingComma": "all"
+}
