+++ apps/landing/components/Hero.tsx
