+++ apps/landing/components/PageIllustration.tsx
