+++ apps/landing/pages/index.tsx
@@ -1,43 +1,38 @@
 import { Inter } from "next/font/google";
 import Head from "next/head";
-import Footer from "../partials/Footer";
-import Header from "../partials/Header";
+import React from "react";
+import Features from "../components/Features";
+import Footer from "../components/Footer";
+import Header from "../components/Header";
+import Hero from "../components/Hero";
 
 const inter = Inter({ subsets: ["latin"] });
 
 export default function Home() {
-	return (
-		<>
-			<Head>
-				<title>Create Next App</title>
-				<meta name="description" content="Generated by create next app" />
-				<meta name="viewport" content="width=device-width, initial-scale=1" />
-				<link rel="icon" href="/favicon.ico" />
-			</Head>
+  return (
+    <>
+      <Head>
+        <title>Create Next App</title>
 
-			<div className="flex flex-col min-h-screen overflow-hidden">
-				{/*  Site header */}
-				<Header />
+        <link rel="icon" href="/favicon.ico" />
+        <link rel="icon" type="image/svg+xml" href="/favicon.svg" />
 
-				{/*  Page content */}
-				<main className="grow">
-					{/*  Page illustration */}
-					<div
-						className="relative max-w-6xl mx-auto h-0 pointer-events-none"
-						aria-hidden="true"
-					>
-						{/* <PageIllustration /> */}
-					</div>
+        <meta name="description" content="Generated by create next app" />
+        <meta name="viewport" content="width=device-width, initial-scale=1.0" />
 
-					{/*  Page sections */}
-					{/* <HeroHome /> */}
-					<div className="hidden">{/* <FeaturesBlocks /> */}</div>
-					{/* <FeaturesZigZag /> */}
-				</main>
+        <title>Embrace Community</title>
+      </Head>
 
-				{/*  Site footer */}
-				<Footer />
-			</div>
-		</>
-	);
+      <div className="flex flex-col min-h-screen overflow-hidden">
+        <Header />
+
+        <main className="grow">
+          <Hero />
+          <Features />
+        </main>
+
+        <Footer />
+      </div>
+    </>
+  );
 }
